created a constant for the template markdown
